+++ Client/SearchPage/ResultComponents/StudyCard.razor.css
@@ -97,14 +97,7 @@
         outline: 0.2rem solid var(--graphic-emph-colour);
     }
 
-/*
-.details-button-text {
-    
-    color: darkslategray;
-    white-space: nowrap;
-    text-decoration: none;
-}
-    */
+
